Adopt JasperFx.Events.ComplianceTests and retire the mirrored Marten test files
Polecat side of marten#5117 (program marten#5119, epic marten#5110). Polecat.Tests
now enrolls the shared cross-store event sourcing compliance suites instead of
carrying hand-mirrored copies of Marten's tests, so the two products' event
sourcing expectations can no longer drift apart.

Added: PolecatComplianceFixture, closing
EventStoreComplianceFixture<TOperations, TQuerySession> over Polecat's
<IDocumentSession, IQuerySession> pair (~150 lines), a ComplianceQuerySession
global alias so the shared self-aggregating fixtures bind EvolveAsync to
Polecat's IQuerySession, and four empty enrollment subclasses. Everything else
comes from the source-only package.

Retired: self_aggregating_evolve_method.cs, assign_tag_where_tests.cs, and the
behavioral half of dcb_tag_query_and_consistency_tests.cs. What survives of the
last one is dcb_documentation_samples.cs, which keeps the twelve
sample_polecat_dcb_* snippet blocks docs/events/dcb.md pulls from plus the
tag/event/aggregate types the other DCB fixtures share. auto_discover_aggregate_types
and projection_sg_dispatch_audit_tests now reference the shared aggregates.

## Product bug the adoption surfaced

QueryByTagsAsync selected e.stream_id and then never mapped it onto the event
envelope -- the reader had a bare `// stream_id at index 2` comment where the
assignment should have been. Every event returned from a DCB tag query carried
StreamId == Guid.Empty. Fixed by hydrating stream identity exactly the way
PolecatEventLoader does, honoring StreamIdentity for the AsString case.

This is precisely why the shared suites keep the strongest assertion: our mirror
of assign_tag_where_by_stream_id had been weakened to check Data types where
Marten asserts e.StreamId == stream1, which hid the bug. The compliance suite
kept Marten's assertion and it failed here on the first run.

## Seam change this drove

Polecat runs its unit of work in parallel and surfaces DcbConcurrencyException
inside an AggregateException, where Marten throws it directly. The failure
semantics are identical, so the shared suites assert through a
ShouldFailWithAsync<T> helper that accepts either shape rather than forcing
either product to change. Recorded as a ninth cross-store drift item.

Polecat.Tests net9.0: 1629 passed, 0 failed, 3 skipped (42/42 compliance).
