+++ src/Polecat.Tests/Compliance/polecat_event_store_compliance.cs
@@ -1,0 +1,23 @@
+using JasperFx.Events.ComplianceTests;
+
+namespace Polecat.Tests.Compliance;
+
+/*
+ * Polecat's enrollment in the cross-store event sourcing compliance suites. Each class below is
+ * empty on purpose: the behavior lives once in JasperFx.Events.ComplianceTests and is closed here
+ * over Polecat's IEventStore<IDocumentSession, IQuerySession> session pair through
+ * PolecatComplianceFixture. Marten enrolls the same way, so drift between the two products' copies
+ * of these tests is no longer possible.
+ */
+
+public class self_aggregating_evolve_compliance
+    : SelfAggregatingEvolveCompliance<PolecatComplianceFixture, IDocumentSession, IQuerySession>;
+
+public class dcb_tag_query_and_consistency_compliance
+    : DcbTagQueryAndConsistencyCompliance<PolecatComplianceFixture, IDocumentSession, IQuerySession>;
+
+public class assign_tag_where_compliance
+    : AssignTagWhereCompliance<PolecatComplianceFixture, IDocumentSession, IQuerySession>;
+
+public class async_daemon_compliance
+    : AsyncDaemonCompliance<PolecatComplianceFixture, IDocumentSession, IQuerySession>;
