--- src/Polecat.Tests/Events/dcb_tag_query_and_consistency_tests.cs
@@ -1,771 +1,0 @@
-#nullable enable
-using JasperFx.Events;
-using JasperFx.Events.Tags;
-using Polecat.Events.Dcb;
-using Polecat.Tests.Harness;
-
-namespace Polecat.Tests.Events;
-
-#region sample_polecat_dcb_tag_type_definitions
-// Strong-typed tag identifiers
-public record StudentId(Guid Value);
-public record CourseId(Guid Value);
-#endregion
-
-#region sample_polecat_dcb_domain_events
-// Domain events
-public record StudentEnrolled(string StudentName, string CourseName);
-public record AssignmentSubmitted(string AssignmentName, int Score);
-public record StudentDropped(string Reason);
-#endregion
-
-// Event with tag-typed properties for inference testing
-public record StudentGraded(StudentId StudentId, CourseId CourseId, int Grade);
-
-// Event with NO tag-typed properties — should fail inference
-public record SystemNotification(string Message);
-
-#region sample_polecat_dcb_aggregate
-// Aggregate for DCB
-public partial class StudentCourseEnrollment
-{
-    public Guid Id { get; set; }
-    public string StudentName { get; set; } = "";
-    public string CourseName { get; set; } = "";
-    public List<string> Assignments { get; set; } = new();
-    public bool IsDropped { get; set; }
-
-    public void Apply(StudentEnrolled e)
-    {
-        StudentName = e.StudentName;
-        CourseName = e.CourseName;
-    }
-
-    public void Apply(AssignmentSubmitted e)
-    {
-        Assignments.Add(e.AssignmentName);
-    }
-
-    public void Apply(StudentDropped e)
-    {
-        IsDropped = true;
-    }
-}
-#endregion
-
-// Polecat always uses Quick append (direct INSERT with OUTPUT seq_id).
-// Tags are inserted immediately after each event, so DCB works with the only append mode available.
-[Collection("integration")]
-public class dcb_tag_query_and_consistency_tests : IntegrationContext
-{
-    public dcb_tag_query_and_consistency_tests(DefaultStoreFixture fixture) : base(fixture)
-    {
-    }
-
-    #region sample_polecat_dcb_registering_tag_types
-    public override async ValueTask InitializeAsync()
-    {
-        await StoreOptions(opts =>
-        {
-            // Register tag types -- each gets its own table (pc_event_tag_student, pc_event_tag_course)
-            opts.Events.RegisterTagType<StudentId>("student")
-                .ForAggregate<StudentCourseEnrollment>();
-            opts.Events.RegisterTagType<CourseId>("course")
-                .ForAggregate<StudentCourseEnrollment>();
-        });
-    }
-    #endregion
-
-    [Fact]
-    public async Task can_query_events_by_single_tag()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        #region sample_polecat_dcb_tagging_events
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-        #endregion
-
-        await using var session2 = theStore.LightweightSession();
-        #region sample_polecat_dcb_query_by_single_tag
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-        #endregion
-
-        events.Count.ShouldBe(1);
-        events[0].Data.ShouldBeOfType<StudentEnrolled>().StudentName.ShouldBe("Alice");
-    }
-
-    [Fact]
-    public async Task can_query_events_by_multiple_tags_with_or()
-    {
-        var student1 = new StudentId(Guid.NewGuid());
-        var student2 = new StudentId(Guid.NewGuid());
-        var course = new CourseId(Guid.NewGuid());
-        var stream1 = Guid.NewGuid();
-        var stream2 = Guid.NewGuid();
-
-        var e1 = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        e1.WithTag(student1, course);
-        theSession.Events.Append(stream1, e1);
-
-        var e2 = theSession.Events.BuildEvent(new StudentEnrolled("Bob", "Math"));
-        e2.WithTag(student2, course);
-        theSession.Events.Append(stream2, e2);
-
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        #region sample_polecat_dcb_query_multiple_tags_or
-        var query = new EventTagQuery()
-            .Or<StudentId>(student1)
-            .Or<StudentId>(student2);
-
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-        #endregion
-        events.Count.ShouldBe(2);
-    }
-
-    [Fact]
-    public async Task can_query_events_across_distinct_tag_types_with_or()
-    {
-        // The core DCB boundary case: events on different streams carry DIFFERENT single tags, and the
-        // query OR-combines distinct tag types. Each matching event carries only one of the queried tag
-        // types — a regression guard against the INNER JOIN bug that required every event to carry all
-        // queried tag types (which collapsed this query to zero rows).
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-
-        // Tagged with ONLY the student
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId);
-        theSession.Events.Append(Guid.NewGuid(), enrolled);
-
-        // Tagged with ONLY the course
-        var submitted = theSession.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        submitted.WithTag(courseId);
-        theSession.Events.Append(Guid.NewGuid(), submitted);
-
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery()
-            .Or<StudentId>(studentId)
-            .Or<CourseId>(courseId);
-
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-        events.Count.ShouldBe(2);
-        events.ShouldContain(e => e.Data is StudentEnrolled);
-        events.ShouldContain(e => e.Data is AssignmentSubmitted);
-    }
-
-    [Fact]
-    public async Task can_query_events_by_tag_with_event_type_filter()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-
-        var submitted = theSession.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        submitted.WithTag(studentId, courseId);
-
-        theSession.Events.Append(streamId, enrolled, submitted);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        #region sample_polecat_dcb_query_by_event_type
-        var query = new EventTagQuery()
-            .Or<AssignmentSubmitted, StudentId>(studentId);
-
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-        #endregion
-        events.Count.ShouldBe(1);
-        events[0].Data.ShouldBeOfType<AssignmentSubmitted>().AssignmentName.ShouldBe("HW1");
-    }
-
-    [Fact]
-    public async Task query_returns_empty_when_no_matching_tags()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var otherStudentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(otherStudentId);
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-        events.Count.ShouldBe(0);
-    }
-
-    [Fact]
-    public async Task can_aggregate_events_by_tags()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-
-        var submitted = theSession.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        submitted.WithTag(studentId, courseId);
-
-        theSession.Events.Append(streamId, enrolled, submitted);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        #region sample_polecat_dcb_aggregate_by_tags
-        var query = new EventTagQuery()
-            .Or<StudentId>(studentId)
-            .Or<CourseId>(courseId);
-
-        var aggregate = await session2.Events.AggregateByTagsAsync<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-        #endregion
-        aggregate.ShouldNotBeNull();
-        aggregate.StudentName.ShouldBe("Alice");
-        aggregate.CourseName.ShouldBe("Math");
-        aggregate.Assignments.ShouldContain("HW1");
-    }
-
-    [Fact]
-    public async Task aggregate_by_tags_returns_null_when_no_events()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var aggregate = await session2.Events.AggregateByTagsAsync<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-        aggregate.ShouldBeNull();
-    }
-
-    [Fact]
-    public async Task can_fetch_for_writing_by_tags_happy_path()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        #region sample_polecat_dcb_fetch_for_writing_by_tags
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session2.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // Read current state
-        var aggregate = boundary.Aggregate; // may be null if no events yet
-        var lastSequence = boundary.LastSeenSequence;
-
-        // Append via boundary
-        var assignment = session2.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        assignment.WithTag(studentId, courseId);
-        boundary.AppendOne(assignment);
-
-        // Save -- will throw DcbConcurrencyException if another session
-        // appended matching events after our read
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-        #endregion
-
-        boundary.Aggregate.ShouldNotBeNull();
-        boundary.Aggregate!.StudentName.ShouldBe("Alice");
-        boundary.Events.Count.ShouldBe(1);
-    }
-
-    [Fact]
-    public async Task fetch_for_writing_appends_to_existing_tag_derived_stream_without_collision()
-    {
-        // Seed the student's stream using the tag value as the stream id, so the boundary's
-        // tag-derived routing targets this PRE-EXISTING stream on save. Before the fix the boundary
-        // used StreamAction.Start here (TryFindStream only sees the current session's pending work),
-        // throwing ExistingStreamIdCollisionException.
-        var studentId = new StudentId(Guid.NewGuid());
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId);
-        theSession.Events.Append(studentId.Value, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session2.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        var submitted = session2.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        submitted.WithTag(studentId);
-        boundary.AppendOne(submitted);
-
-        await Should.NotThrowAsync(async () => await session2.SaveChangesAsync());
-
-        await using var session3 = theStore.LightweightSession();
-        var events = await session3.Events.QueryByTagsAsync(new EventTagQuery().Or<StudentId>(studentId), TestContext.Current.CancellationToken);
-        events.ShouldContain(e => e.Data is AssignmentSubmitted);
-    }
-
-    [Fact]
-    public async Task fetch_for_writing_by_tags_detects_concurrency_violation()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1: fetch for writing
-        await using var session1 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session1.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // Session 2: append a conflicting event BEFORE session 1 saves
-        await using var session2 = theStore.LightweightSession();
-        var conflicting = session2.Events.BuildEvent(new AssignmentSubmitted("HW-conflict", 50));
-        conflicting.WithTag(studentId, courseId);
-        session2.Events.Append(streamId, conflicting);
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1: try to save — should throw DcbConcurrencyException
-        var assignment = session1.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        assignment.WithTag(studentId, courseId);
-        boundary.AppendOne(assignment);
-
-        #region sample_polecat_dcb_handling_concurrency
-        try
-        {
-            await session1.SaveChangesAsync(TestContext.Current.CancellationToken);
-        }
-        catch (AggregateException ex) when (ex.InnerExceptions.OfType<DcbConcurrencyException>().Any())
-        {
-            // Reload and retry -- the boundary's tag query had new matching events
-            var violation = ex.InnerExceptions.OfType<DcbConcurrencyException>().First();
-            // violation.Query -- the original tag query
-            // violation.LastSeenSequence -- the sequence at time of read
-        }
-        #endregion
-    }
-
-    [Fact]
-    public async Task fetch_for_writing_by_tags_no_violation_when_unrelated_events_appended()
-    {
-        var student1 = new StudentId(Guid.NewGuid());
-        var student2 = new StudentId(Guid.NewGuid());
-        var course = new CourseId(Guid.NewGuid());
-        var stream1 = Guid.NewGuid();
-        var stream2 = Guid.NewGuid();
-
-        var enrolled1 = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled1.WithTag(student1, course);
-        theSession.Events.Append(stream1, enrolled1);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1: fetch for writing for student1
-        await using var session1 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(student1);
-        var boundary = await session1.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // Session 2: append event for DIFFERENT student — should NOT conflict
-        await using var session2 = theStore.LightweightSession();
-        var enrolled2 = session2.Events.BuildEvent(new StudentEnrolled("Bob", "Math"));
-        enrolled2.WithTag(student2, course);
-        session2.Events.Append(stream2, enrolled2);
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1: save should succeed
-        var assignment = session1.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        assignment.WithTag(student1, course);
-        boundary.AppendOne(assignment);
-
-        await session1.SaveChangesAsync(TestContext.Current.CancellationToken); // Should not throw
-    }
-
-    [Fact]
-    public async Task events_across_multiple_streams_can_be_queried_by_tag()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var course1 = new CourseId(Guid.NewGuid());
-        var course2 = new CourseId(Guid.NewGuid());
-        var stream1 = Guid.NewGuid();
-        var stream2 = Guid.NewGuid();
-
-        var enrolled1 = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled1.WithTag(studentId, course1);
-        theSession.Events.Append(stream1, enrolled1);
-
-        var enrolled2 = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Science"));
-        enrolled2.WithTag(studentId, course2);
-        theSession.Events.Append(stream2, enrolled2);
-
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-
-        events.Count.ShouldBe(2);
-    }
-
-    [Fact]
-    public async Task query_events_ordered_by_sequence()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-
-        var hw1 = theSession.Events.BuildEvent(new AssignmentSubmitted("HW1", 90));
-        hw1.WithTag(studentId, courseId);
-
-        var hw2 = theSession.Events.BuildEvent(new AssignmentSubmitted("HW2", 85));
-        hw2.WithTag(studentId, courseId);
-
-        theSession.Events.Append(streamId, enrolled, hw1, hw2);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var events = await session2.Events.QueryByTagsAsync(query, TestContext.Current.CancellationToken);
-
-        events.Count.ShouldBe(3);
-        events[0].Sequence.ShouldBeLessThan(events[1].Sequence);
-        events[1].Sequence.ShouldBeLessThan(events[2].Sequence);
-    }
-
-    [Fact]
-    public async Task fetch_for_writing_with_empty_result_still_enforces_consistency()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-
-        // Fetch for writing when no events exist
-        await using var session1 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session1.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        boundary.Aggregate.ShouldBeNull();
-        boundary.Events.Count.ShouldBe(0);
-        boundary.LastSeenSequence.ShouldBe(0);
-
-        // Another session appends a matching event before save
-        await using var session2 = theStore.LightweightSession();
-        var enrolled = session2.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        var streamId = Guid.NewGuid();
-        session2.Events.Append(streamId, enrolled);
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1 tries to save — should detect the new matching event
-        var e = session1.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        e.WithTag(studentId, courseId);
-        boundary.AppendOne(e);
-
-        var ex = await Should.ThrowAsync<AggregateException>(async () =>
-        {
-            await session1.SaveChangesAsync();
-        });
-        ex.InnerExceptions.ShouldContain(e => e is DcbConcurrencyException);
-    }
-
-    [Fact]
-    public async Task can_fetch_for_writing_by_tags_via_batch_query()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var batch = session2.CreateBatchQuery();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundaryTask = batch.FetchForWritingByTags<StudentCourseEnrollment>(query);
-        await batch.Execute(TestContext.Current.CancellationToken);
-
-        var boundary = await boundaryTask;
-        boundary.Aggregate.ShouldNotBeNull();
-        boundary.Aggregate!.StudentName.ShouldBe("Alice");
-        boundary.Events.Count.ShouldBe(1);
-        boundary.LastSeenSequence.ShouldBeGreaterThan(0);
-
-        // Append via boundary and save
-        var assignment = session2.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        assignment.WithTag(studentId, courseId);
-        boundary.AppendOne(assignment);
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-    }
-
-    [Fact]
-    public async Task batch_query_fetch_for_writing_by_tags_detects_concurrency_violation()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1: fetch via batch query
-        await using var session1 = theStore.LightweightSession();
-        var batch = session1.CreateBatchQuery();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundaryTask = batch.FetchForWritingByTags<StudentCourseEnrollment>(query);
-        await batch.Execute(TestContext.Current.CancellationToken);
-        var boundary = await boundaryTask;
-
-        // Session 2: append conflicting event
-        await using var session2 = theStore.LightweightSession();
-        var conflicting = session2.Events.BuildEvent(new AssignmentSubmitted("HW-conflict", 50));
-        conflicting.WithTag(studentId, courseId);
-        session2.Events.Append(streamId, conflicting);
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Session 1: try to save — should throw
-        var assignment = session1.Events.BuildEvent(new AssignmentSubmitted("HW1", 95));
-        assignment.WithTag(studentId, courseId);
-        boundary.AppendOne(assignment);
-
-        var ex = await Should.ThrowAsync<AggregateException>(async () =>
-        {
-            await session1.SaveChangesAsync();
-        });
-        ex.InnerExceptions.ShouldContain(e => e is DcbConcurrencyException);
-    }
-
-    #region sample_polecat_dcb_events_exist_async
-    [Fact]
-    public async Task events_exist_returns_true_when_matching_events_found()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Check existence -- lightweight, no event loading
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var exists = await session2.Events.EventsExistAsync(query, TestContext.Current.CancellationToken);
-        exists.ShouldBeTrue();
-    }
-    #endregion
-
-    [Fact]
-    public async Task events_exist_returns_false_when_no_matching_events()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var exists = await session2.Events.EventsExistAsync(query, TestContext.Current.CancellationToken);
-        exists.ShouldBeFalse();
-    }
-
-    [Fact]
-    public async Task events_exist_with_event_type_filter()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-
-        // Should find StudentEnrolled
-        var query1 = new EventTagQuery().Or<StudentEnrolled, StudentId>(studentId);
-        (await session2.Events.EventsExistAsync(query1, TestContext.Current.CancellationToken)).ShouldBeTrue();
-
-        // Should NOT find AssignmentSubmitted (none appended)
-        var query2 = new EventTagQuery().Or<AssignmentSubmitted, StudentId>(studentId);
-        (await session2.Events.EventsExistAsync(query2, TestContext.Current.CancellationToken)).ShouldBeFalse();
-    }
-
-    [Fact]
-    public async Task events_exist_via_batch_query_positive()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var batch = session2.CreateBatchQuery();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var existsTask = batch.EventsExist(query);
-        await batch.Execute(TestContext.Current.CancellationToken);
-
-        (await existsTask).ShouldBeTrue();
-    }
-
-    [Fact]
-    public async Task events_exist_via_batch_query_negative()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-
-        await using var session2 = theStore.LightweightSession();
-        var batch = session2.CreateBatchQuery();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var existsTask = batch.EventsExist(query);
-        await batch.Execute(TestContext.Current.CancellationToken);
-
-        (await existsTask).ShouldBeFalse();
-    }
-
-    [Fact]
-    public async Task fetch_for_writing_by_tags_throws_on_empty_query()
-    {
-        var query = new EventTagQuery();
-        await Should.ThrowAsync<ArgumentException>(async () =>
-        {
-            await theSession.Events.FetchForWritingByTags<StudentCourseEnrollment>(query);
-        });
-    }
-
-    [Fact]
-    public async Task append_event_with_inferred_tags_from_properties()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        // Seed initial event with explicit tags
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Fetch for writing
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session2.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // Append a raw event that has StudentId and CourseId properties —
-        // tags should be inferred automatically
-        boundary.AppendOne(new StudentGraded(studentId, courseId, 95));
-
-        // Should succeed — tags inferred from properties
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Verify the event is discoverable by tag query
-        await using var session3 = theStore.LightweightSession();
-        var events = await session3.Events.QueryByTagsAsync(
-            new EventTagQuery().Or<StudentId>(studentId), TestContext.Current.CancellationToken);
-        events.Count.ShouldBe(2);
-        events[1].Data.ShouldBeOfType<StudentGraded>().Grade.ShouldBe(95);
-    }
-
-    [Fact]
-    public async Task append_event_with_no_tags_and_no_inferable_properties_throws()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        // Seed initial event
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Fetch for writing
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session2.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // Append an event with no tags and no tag-typed properties — should throw
-        Should.Throw<InvalidOperationException>(() =>
-        {
-            boundary.AppendOne(new SystemNotification("test"));
-        });
-    }
-
-    [Fact]
-    public async Task append_already_wrapped_event_with_explicit_tags_works()
-    {
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        // Seed initial event
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        // Fetch for writing
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session2.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // Append an already-wrapped event with explicit tags
-        var graded = session2.Events.BuildEvent(new StudentGraded(studentId, courseId, 88));
-        graded.WithTag(studentId, courseId);
-        boundary.AppendOne(graded);
-
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-    }
-
-    [Fact]
-    public async Task append_event_with_tag_having_no_aggregate_type_creates_new_stream()
-    {
-        // Register a tag type WITHOUT an aggregate association
-        await StoreOptions(opts =>
-        {
-            opts.Events.RegisterTagType<StudentId>("student");
-            // CourseId registered WITHOUT ForAggregate
-            opts.Events.RegisterTagType<CourseId>("course");
-        });
-
-        var studentId = new StudentId(Guid.NewGuid());
-        var courseId = new CourseId(Guid.NewGuid());
-        var streamId = Guid.NewGuid();
-
-        var enrolled = theSession.Events.BuildEvent(new StudentEnrolled("Alice", "Math"));
-        enrolled.WithTag(studentId, courseId);
-        theSession.Events.Append(streamId, enrolled);
-        await theSession.SaveChangesAsync(TestContext.Current.CancellationToken);
-
-        await using var session2 = theStore.LightweightSession();
-        var query = new EventTagQuery().Or<StudentId>(studentId);
-        var boundary = await session2.Events.FetchForWritingByTags<StudentCourseEnrollment>(query, TestContext.Current.CancellationToken);
-
-        // CourseId tag has no AggregateType — should create a new stream per event
-        var graded = session2.Events.BuildEvent(new StudentGraded(studentId, courseId, 90));
-        graded.WithTag(courseId);
-        boundary.AppendOne(graded);
-
-        // Should succeed — unrouted tag creates a new stream
-        await session2.SaveChangesAsync(TestContext.Current.CancellationToken);
-    }
-}
